# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: InterviewAce API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /discuss/posts:
    get:
      operationId: listDiscussPosts
      tags: [discuss]
      summary: List community discuss posts
      parameters:
        - name: topicTag
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscussPost"
    post:
      operationId: createDiscussPost
      tags: [discuss]
      summary: Create a discuss post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscussPostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscussPost"
  /discuss/posts/{id}:
    get:
      operationId: getDiscussPost
      tags: [discuss]
      summary: Get a discuss post with comments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscussPostDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteDiscussPost
      tags: [discuss]
      summary: Delete own discuss post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /discuss/posts/{id}/comments:
    post:
      operationId: createPostComment
      tags: [discuss]
      summary: Add a comment to a discuss post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostCommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostComment"
        "404":
          description: Not found
  /discuss/posts/{id}/upvote:
    post:
      operationId: toggleDiscussPostUpvote
      tags: [discuss]
      summary: Toggle upvote on a discuss post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscussPost"
        "404":
          description: Not found

  # ---------- Matching ----------
  /discuss/match:
    post:
      operationId: createMatchRequest
      tags: [matching]
      summary: Request an anonymous co-learner match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchRequestInput"
      responses:
        "201":
          description: Match request created (may already be matched)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchRequestStatus"
  /discuss/match/{id}:
    get:
      operationId: getMatchRequest
      tags: [matching]
      summary: Poll a match request's status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match request status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchRequestStatus"
        "404":
          description: Not found
    delete:
      operationId: cancelMatchRequest
      tags: [matching]
      summary: Cancel a pending match request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
        "404":
          description: Not found

  # ---------- Sessions ----------
...
components:
  schemas:
...
    DiscussPost:
      type: object
      required:
        [id, authorHandle, avatarColor, type, title, content, topicTag, upvoteCount, commentCount, isUpvotedByMe, isMine, createdAt]
      properties:
        id:
          type: integer
        authorHandle:
          type: string
        avatarColor:
          type: string
        type:
          type: string
          enum: [question, discussion, resource]
        title:
          type: string
        content:
          type: string
        topicTag:
          type: ["string", "null"]
        upvoteCount:
          type: integer
        commentCount:
          type: integer
        isUpvotedByMe:
          type: boolean
        isMine:
          type: boolean
        createdAt:
          type: string
          format: date-time
    DiscussPostInput:
      type: object
      required: [type, title, content]
      properties:
        type:
          type: string
          enum: [question, discussion, resource]
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        topicTag:
          type: string
    PostComment:
      type: object
      required: [id, postId, authorHandle, avatarColor, content, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorHandle:
          type: string
        avatarColor:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    PostCommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
    DiscussPostDetail:
      type: object
      required: [post, comments]
      properties:
        post:
          $ref: "#/components/schemas/DiscussPost"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/PostComment"

    # ---------- Matching ----------
    MatchRequestInput:
      type: object
      required: [topic, skillLevel, chatType, durationMinutes]
      properties:
        topic:
          type: string
        skillLevel:
          type: string
          enum: [beginner, intermediate, advanced, any]
        chatType:
          type: string
          enum: [text, voice]
        durationMinutes:
          type: integer
        language:
          type: string
        anonymous:
          type: boolean
    MatchRequestStatus:
      type: object
      required: [id, status, topic, chatType, durationMinutes, createdAt]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [waiting, matched, cancelled, expired]
        topic:
          type: string
        chatType:
          type: string
          enum: [text, voice]
        durationMinutes:
          type: integer
        sessionId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time

    # ---------- Sessions ----------
...