# lib/api-zod/src/generated/types/matchRequestInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * InterviewAce API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatchRequestInputChatType } from './matchRequestInputChatType';
import type { MatchRequestInputSkillLevel } from './matchRequestInputSkillLevel';

export interface MatchRequestInput {
  topic: string;
  skillLevel: MatchRequestInputSkillLevel;
  chatType: MatchRequestInputChatType;
  durationMinutes: number;
  language?: string;
  anonymous?: boolean;
}